+++ tools/virtio/include/virtio.h
@@ -223,7 +223,7 @@ void update_used_ring(VirtQueue *vq, uint16_t idx, uint32_t iolen);
 // Batch version of update_used_ring: writes `count` used-ring entries
 // with a single pair of write barriers instead of two per entry.
 void update_used_ring_batch(VirtQueue *vq, const uint16_t *indices,
-                            const uint32_t *lens, int count);
+                            const uint32_t *lens, size_t count);
 
 uint64_t virtio_mmio_read(VirtIODevice *vdev, uint64_t offset, unsigned size);
